pyproject.toml
[build-system]
requires = ['setuptools>=60', 'wheel']
build-backend = 'setuptools.build_meta'

[project]
name = 'lattifai'
version = '0.4.5'
description = "Lattifai Python SDK: Seamless Integration with Lattifai's Speech and Video AI Services"
keywords = ['lattifai', 'speech recognition', 'video analysis', 'ai', 'sdk', 'api client']
readme = 'README.md'
license = {file = 'LICENSE'}
authors = [
  { name = 'Lattifai Technologies', email = 'tech@lattifai.com' },
]
maintainers = [
  { name = 'Lattice', email = 'tech@lattifai.com' },
]
requires-python = '>=3.10,<3.14'

dependencies = [
    "lattifai-core>=0.2.1",
    "httpx",
    "python-dotenv",
    "lhotse>=1.26.0",
    "colorful>=0.5.6",
    "pysubs2",
    "praatio",
    "tgt",
    "onnxruntime",
    "g2p-phonemizer==0.1.1",
    "wtpsplit>=2.1.6",
    "av",
    # agentic workflow dependencies
    "questionary>=2.0",
    "yt-dlp",
    "pycryptodome",
    "google-genai"
]

classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Intended Audience :: Developers',
    'Intended Audience :: Science/Research',
    'License :: OSI Approved :: Apache Software License',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Programming Language :: Python :: 3.13',
    'Operating System :: MacOS :: MacOS X',
    'Operating System :: POSIX :: Linux',
    'Operating System :: Microsoft :: Windows',
    'Topic :: Multimedia :: Sound/Audio',
    'Topic :: Multimedia :: Video',
    'Topic :: Scientific/Engineering :: Artificial Intelligence',
]

[project.urls]
Homepage = 'https://github.com/lattifai/lattifai-python'
Documentation = 'https://github.com/lattifai/lattifai-python/README.md'
'Bug Tracker' = 'https://github.com/lattifai/lattifai-python/issues'
Discussions = 'https://github.com/lattifai/lattifai-python/discussions'
Changelog = 'https://github.com/lattifai/lattifai-python/CHANGELOG.md'

[project.scripts]
lattifai = 'lattifai.bin:cli'
lai = 'lattifai.bin:cli'

[tool.setuptools]
packages = ["lattifai", "lattifai.io", "lattifai.tokenizer", "lattifai.workers", "lattifai.bin", "lattifai.workflows", "lattifai.workflows.prompts"]

[tool.setuptools.package-dir]
lattifai = "src/lattifai"

[tool.setuptools.package-data]
"lattifai.workflows.prompts" = ["**/*.txt", "**/*.md"]

[project.optional-dependencies]
numpy = ['numpy']
test = ['pytest', 'pytest-cov', 'pytest-asyncio', 'numpy']
all = [
    "numpy",
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
]

[tool.pytest.ini_options]
addopts = "--durations=10"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
]
